stock-research-app/api/requirements.txt
azure-functions==1.20.0
azure-functions-durable==1.2.9
azure-identity==1.17.1
azure-cosmos==4.7.0
azure-storage-blob==12.22.0
openai>=1.40.0
httpx==0.27.2
markdown-it-py==3.0.0
reportlab==4.2.2
pydantic==2.8.2
azure-communication-email==1.0.0
tenacity==8.5.0
readability-lxml==0.8.1
beautifulsoup4==4.12.3
lxml==5.2.2
